Move to recorder/util.py
